+++ src/routes/IndexPage/views/CurrencyTable/index.ts
@@ -1,3 +1,4 @@
 import CurrencyTable from './CurrencyTable';
 
+export * from './CurrencyTable';
 export default CurrencyTable;
