+++ src/shared/Header/Header.tsx
@@ -4,7 +4,7 @@ import {Button} from '@steroidsjs/core/ui/form';
 import {useBem, useDispatch, useSelector} from '@steroidsjs/core/hooks';
 import getCurrencyList from 'store/actions/currency';
 import {selectCurrencyList, selectCurrencyListIsLoading} from 'store/reducers/currency';
-import {DEFAULT_QUERY_PARAMS} from '../../routes/IndexPage/views/CurrencyTable/CurrencyTable';
+import {DEFAULT_QUERY_PARAMS} from '../../routes/IndexPage/views/CurrencyTable';
 
 const MILLISECOND = 1000;
